INT-2730 Move OrderlyShutDownAware Interface
Move from core to context.
