+++ spring-integration-jmx/src/main/java/org/springframework/integration/monitor/IntegrationMBeanExporter.java
@@ -58,9 +58,9 @@ import org.springframework.core.task.support.ExecutorServiceAdapter;
 import org.springframework.integration.MessageChannel;
 import org.springframework.integration.MessagingException;
 import org.springframework.integration.channel.QueueChannel;
+import org.springframework.integration.context.OrderlyShutdownCapable;
 import org.springframework.integration.core.MessageHandler;
 import org.springframework.integration.core.MessageSource;
-import org.springframework.integration.core.OrderlyShutdownCapable;
 import org.springframework.integration.core.PollableChannel;
 import org.springframework.integration.endpoint.AbstractEndpoint;
 import org.springframework.jmx.export.MBeanExporter;
